requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
boto3==1.40.0
botocore==1.40.0
azure-identity==1.23.0
azure-mgmt-resource==23.3.0
azure-keyvault-secrets==4.10.0
pysnmp==7.1.17
python-dotenv==1.1.1
SQLAlchemy==2.0.43
alembic==1.16.5
APScheduler==3.11.0
python-jose[cryptography]==3.5.0
Authlib==1.6.1
itsdangerous==2.2.0
httpx==0.28.1
